# apps/api/openapi/libif-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/taxonomy/tags": {
      "get": {
        "operationId": "listTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaxonomyTagDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "List tag options for staff document metadata.",
        "tags": [
          "Taxonomy"
        ]
      }
    },
    "/api/taxonomy/tags/{id}/impact": {
      "get": {
        "operationId": "getTagImpact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyTagImpactDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Get tag usage impact metrics.",
        "tags": [
          "Taxonomy"
        ]
      }
    },
...
    "/api/admin/tags/{id}/impact": {
      "get": {
        "operationId": "getTagImpact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyTagImpactDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Get tag usage impact metrics.",
        "tags": [
          "Admin Taxonomy"
        ]
      }
    },
    "/api/admin/tags": {
      "post": {
        "operationId": "createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaxonomyTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyTagDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Create a tag.",
        "tags": [
          "Admin Taxonomy"
        ]
      }
    },
    "/api/admin/tags/{id}": {
      "patch": {
        "operationId": "updateTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaxonomyTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyTagDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Edit a tag.",
        "tags": [
          "Admin Taxonomy"
        ]
      },
      "delete": {
        "operationId": "deleteTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag deleted successfully."
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Delete a tag and detach it from all documents.",
        "tags": [
          "Admin Taxonomy"
        ]
      }
    },
    "/api/admin/tags/{id}/merge": {
      "post": {
        "operationId": "mergeTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag merged successfully."
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Merge source tag into target tag and delete source tag.",
        "tags": [
          "Admin Taxonomy"
        ]
      }
    },
...
  },
  "info": {
    "title": "LIBIF API",
    "description": "OpenAPI contract for LIBIF web and API integration.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AuthErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "AUTHENTICATION_REQUIRED"
          },
          "message": {
            "type": "string",
            "example": "Authentication is required."
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {}
          },
          "traceId": {
            "type": "string",
            "example": "request-trace-id"
          },
          "status": {
            "type": "number",
            "example": 403
          }
        },
        "required": [
          "code",
          "message",
          "fieldErrors",
          "traceId",
          "status"
        ]
      },
...
      "TaxonomyTagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "TaxonomyTagImpactDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "documentCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "documentCount"
        ]
      },
...
      "CreateTaxonomyTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Digital Preservation"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTaxonomyTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Digital Preservation"
          }
        }
      },
      "MergeTagDto": {
        "type": "object",
        "properties": {
          "targetTagId": {
            "type": "string",
            "description": "Target tag ID to merge source tag into."
          }
        },
        "required": [
          "targetTagId"
        ]
      },
...
    }
  }
}